+++ src/app/collective-rewards/components/CollectiveRewardsChart/index.ts
@@ -1,0 +1,1 @@
+export * from './CollectiveRewardsChart'
