--- src/app/collective-rewards/components/DualAxisChart/index.ts
@@ -1,1 +1,0 @@
-export * from './CollectiveRewardsDualAxisChart'
